backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# Database
supabase==2.3.0
postgrest==0.13.2

# Reddit Discovery
praw==7.7.1

# LLM Providers
groq==0.4.1
google-generativeai==0.3.2

# SEO & Keywords
pytrends==4.9.2

# PDF Generation (Phase 1)
reportlab==4.0.9
# weasyprint==60.2  # Alternative PDF generator

# Image Processing (Phase 1)
Pillow==10.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing FastAPI

# Utilities
python-dateutil==2.8.2
